uiux: properly determine line height and use it for drawing cursor
line height was naively determined to just be the text size. this made
letters like 'p' and 'g' seem as though they were cut off when the
cursor was on them.

this refactor takes into account the ascent and descent of the dummy
space character that is shaped so that lines have a proper height. as a
side bonus, it seems line padding is no longer needed by default
(although it can be worked back in later, separate from the text
layout).