refactor: optimize drawing editor text
`widget.Label` was used within the editor for drawing text segments and
line numbers. it did the job, but it's meant for general purpose labels
that might (for example) be aligned start, middle or end, or that may
have more than one line. a lot of code is required to get this right.

this change adds a `drawText` function which simply draws text, assuming
one line only and not taking any alignment into account. this removes
the need for `widget.Label` within editor.