uiux: maintain relative cursor position if possible when moving by line
previously, moving up or down by line would the cursor to the beginning
of the new line. with this change, the cursor column position will be
maintained if it can be while the user is moving lines.

moving to the end of a line will set this 'preferred' (for lack of a
better term) column index to dynamically be the end of whatever line the
user moves to next.